+++ tests/testthat/test-epinow.R
@@ -18,7 +18,7 @@ expected_out <- c("estimates", "estimated_reported_cases", "summary", "plots", "
 
 test_that("epinow produces expected output when run with default settings", {
   out <- suppressWarnings(epinow(
-    reported_cases = reported_cases,
+    data = reported_cases,
     generation_time = generation_time_opts(example_generation_time),
     delays = delay_opts(c(example_incubation_period, reporting_delay)),
     stan = stan_opts(
@@ -43,7 +43,7 @@ test_that("epinow produces expected output when run with the
   skip_on_os("windows")
   output <- capture.output(suppressMessages(suppressWarnings(
     out <- epinow(
-      reported_cases = reported_cases,
+      data = reported_cases,
       generation_time = generation_time_opts(example_generation_time),
       delays = delay_opts(example_incubation_period + reporting_delay),
       stan = stan_opts(backend = "cmdstanr"),
@@ -67,7 +67,7 @@ test_that("epinow produces expected output when run with the
   skip_on_os("windows")
   output <- capture.output(suppressMessages(suppressWarnings(
     out <- epinow(
-      reported_cases = reported_cases,
+      data = reported_cases,
       generation_time = generation_time_opts(example_generation_time),
       delays = delay_opts(example_incubation_period + reporting_delay),
       stan = stan_opts(method = "laplace", backend = "cmdstanr"),
@@ -90,7 +90,7 @@ test_that("epinow produces expected output when run with the
   skip_on_os("windows")
   output <- capture.output(suppressMessages(suppressWarnings(
     out <- epinow(
-      reported_cases = reported_cases,
+      data = reported_cases,
       generation_time = generation_time_opts(example_generation_time),
       delays = delay_opts(example_incubation_period + reporting_delay),
       stan = stan_opts(method = "pathfinder", backend = "cmdstanr"),
@@ -110,7 +110,7 @@ test_that("epinow produces expected output when run with the
 
 test_that("epinow runs without error when saving to disk", {
   expect_null(suppressWarnings(epinow(
-    reported_cases = reported_cases,
+    data = reported_cases,
     generation_time = generation_time_opts(example_generation_time),
     delays = delay_opts(example_incubation_period + reporting_delay),
     stan = stan_opts(
@@ -124,7 +124,7 @@ test_that("epinow runs without error when saving to disk", {
 
 test_that("epinow can produce partial output as specified", {
   out <- suppressWarnings(epinow(
-    reported_cases = reported_cases,
+    data = reported_cases,
     generation_time = generation_time_opts(
       example_generation_time, weight_prior = FALSE
     ),
@@ -149,7 +149,7 @@ test_that("epinow can produce partial output as specified", {
 
 test_that("epinow fails as expected when given a short timeout", {
   expect_error(suppressWarnings(x = epinow(
-    reported_cases = reported_cases,
+    data = reported_cases,
     generation_time = generation_time_opts(example_generation_time),
     delays = delay_opts(example_incubation_period + reporting_delay),
     stan = stan_opts(
@@ -165,7 +165,7 @@ test_that("epinow fails as expected when given a short timeout", {
 
 test_that("epinow fails if given NUTs arguments when using variational inference", {
   expect_error(suppressWarnings(epinow(
-    reported_cases = reported_cases,
+    data = reported_cases,
     generation_time = generation_time_opts(example_generation_time),
     delays = delay_opts(example_incubation_period + reporting_delay),
     stan = stan_opts(
@@ -180,10 +180,16 @@ test_that("epinow fails if given NUTs arguments when using variational inference
 
 test_that("epinow fails if given variational inference arguments when using NUTs", {
   expect_error(suppressWarnings(epinow(
-    reported_cases = reported_cases,
+    data = reported_cases,
     generation_time = generation_time_opts(example_generation_time),
     delays = delay_opts(example_incubation_period + reporting_delay),
     stan = stan_opts(method = "sampling", tol_rel_obj = 1),
     logs = NULL, verbose = FALSE
   )))
+})
+
+test_that("deprecated arguments are recognised", {
+  expect_deprecated(epinow(reported_cases = reported_cases,
+                      generation_time = generation_time_opts(Fixed(1))
+  ))
 })